[design] userの詳細情報表示ページのデザインを変更
