+++ frontend/components/side-bar.vue
@@ -7,15 +7,25 @@
           </nuxt-link>
         </el-menu-item>
         <el-menu-item index="2">
-          <i class="icon el-icon-user-solid"></i>
+          <nuxt-link to="/users">
+            <i class="icon el-icon-user-solid"></i>
+          </nuxt-link>
         </el-menu-item>
         <el-menu-item index="3">
           <nuxt-link to="/posts/new">
+            <i class="icon el-icon-circle-plus-outline"></i>
+          </nuxt-link>
+        </el-menu-item>
+        <el-menu-item index="4">
+          <nuxt-link to="/posts/search">
             <i class="icon el-icon-search"></i>
           </nuxt-link>
         </el-menu-item>
-        <el-menu-item index="4"><i class="icon el-icon-message-solid"></i></el-menu-item>
-        <el-menu-item index="5"><i class="icon el-icon-info"></i></el-menu-item>
+        <el-menu-item index="5">
+          <nuxt-link :to="{ path: `/users/${$store.state.auth.currentUser.id}` }">
+            <i class="icon el-icon-setting"></i>
+          </nuxt-link>
+        </el-menu-item>
       </el-menu>
     </el-aside>
 </template>
