+++ frontend/nuxt.config.js
@@ -39,8 +39,8 @@ const config = {
     '@/plugins/element-ui',
     { src: '~/plugins/axios.js', ssr: false },
     { src: "~plugins/vue2-google-maps.js" },
-    { src: '@/plugins/localStorage', ssr: false },
     { src: '~/plugins/cookie-storage.js' },
+    { src: '@/plugins/localStorage', ssr: false },
   ],
   /*
   ** Nuxt.js dev-modules
