+++ frontend/pages/users/index.vue
@@ -1,0 +1,3 @@
+<template>
+  <h2>users index</h2>
+</template>
