+++ frontend/plugins/localStorage.js
@@ -1,11 +1,9 @@
 import createPersistedState from 'vuex-persistedstate'
  
 export default ({ store }) => {
- 
     window.onNuxtReady(() => {
         createPersistedState({
-            key: 'dishare-key',
-            storage: window.sessionStorage
+            key: 'dishare-key'
         })(store)
     })
 }
