+++ frontend/store/auth.js
@@ -35,7 +35,7 @@ export const actions = {
       .then( res => {
         commit('setCurrentUser', res.data)
         commit('setIsLoggedIn', true)
-        this.$router.push("/users/mypage")
+        this.$router.push(`/users/${res.data.id}`)
         return res
       }, err => {
         console.log(err)
@@ -51,7 +51,9 @@ export const actions = {
     }).then( res => {
         commit('setCurrentUser', res.data)
         commit('setIsLoggedIn', true)
-        this.$router.push("/users/mypage")
+        console.log('ログインしたよｘ')
+        console.log(res)
+        this.$router.push(`/users/${res.data.id}`)
         return res
       }, err => {
         console.log('ログイン失敗しちゃった。' + ' /stores/auth.js')
