+++ frontend/store/index.js
@@ -1,0 +1,10 @@
+export const state = () => ({
+  // クリックしたUserを詳細表示するためのデータ
+  user: {}
+})
+
+export const mutations = {
+  setUser: (state, userData) => {
+    state.user = userData
+  }
+}
